+++ src/utils/code_highlight/highlight_bat.c
@@ -97,7 +97,7 @@ char* highlight_bat(const char* code) {
                 // Read variable
                 int token_index = 0;
                 while (*p && *p != '%') {
-                    if (token_index < sizeof(token) - 1) {
+                    if (token_index < (int)(sizeof(token) - 1)) {
                         token[token_index++] = *p++;
                     }
                 }
@@ -120,7 +120,7 @@ char* highlight_bat(const char* code) {
                 // Handle keywords or commands
                 int token_index = 0;
                 while (isalnum(*p) || *p == '_') {
-                    if (token_index < sizeof(token) - 1) {
+                    if (token_index < (int)(sizeof(token) - 1)) {
                         token[token_index++] = *p++;
                     } else {
                         p++;
@@ -140,7 +140,7 @@ char* highlight_bat(const char* code) {
                 // Handle numbers
                 int token_index = 0;
                 while (isdigit(*p) || *p == '.') {
-                    if (token_index < sizeof(token) - 1) {
+                    if (token_index < (int)(sizeof(token) - 1)) {
                         token[token_index++] = *p++;
                     } else {
                         p++;
