+++ src/utils/code_highlight/highlight_prolog.c
@@ -132,7 +132,7 @@ char *highlight_prolog(const char *code) {
         // Variables (start with uppercase letter or _)
         int token_index = 0;
         while (isalnum(*p) || *p == '_') {
-          if (token_index < sizeof(token) - 1) {
+          if (token_index < (int)(sizeof(token) - 1)) {
             token[token_index++] = *p++;
           } else {
             p++;
@@ -150,7 +150,7 @@ char *highlight_prolog(const char *code) {
         // Atoms or predicates (start with lowercase letter)
         int token_index = 0;
         while (isalnum(*p) || *p == '_') {
-          if (token_index < sizeof(token) - 1) {
+          if (token_index < (int)(sizeof(token) - 1)) {
             token[token_index++] = *p++;
           } else {
             p++;
@@ -175,7 +175,7 @@ char *highlight_prolog(const char *code) {
         // Handle numbers
         int token_index = 0;
         while (isdigit(*p) || *p == '.') {
-          if (token_index < sizeof(token) - 1) {
+          if (token_index < (int)(sizeof(token) - 1)) {
             token[token_index++] = *p++;
           } else {
             p++;
